+++ src/endpoint/mod.rs
@@ -11,7 +11,6 @@ pub(crate) mod path;
 pub(crate) mod result;
 
 pub(crate) mod and_then;
-pub(crate) mod from_err;
 pub(crate) mod inspect;
 pub(crate) mod join;
 pub(crate) mod join_all;
@@ -36,7 +35,6 @@ pub use self::path::{path, paths};
 pub use self::result::{err, ok, result, EndpointErr, EndpointOk, EndpointResult};
 
 pub use self::and_then::AndThen;
-pub use self::from_err::FromErr;
 pub use self::inspect::Inspect;
 pub use self::join::{Join, Join3, Join4, Join5, Join6};
 pub use self::join_all::{join_all, JoinAll};
